requirements.txt
flask==3.0.3
pymysql==1.1.1
requests==2.32.3
celery==5.4.0
mysql.connector
flask_restful
